Extract supported encodings provider to a trait for re-use with other escaper tests and fix tests that don't assert anything
