+++ test/Helper/EscapeCssTest.php
@@ -16,17 +16,7 @@ use stdClass;
 
 class EscapeCssTest extends TestCase
 {
-    protected $supportedEncodings = [
-        'iso-8859-1',   'iso8859-1',    'iso-8859-5',   'iso8859-5',
-        'iso-8859-15',  'iso8859-15',   'utf-8',        'cp866',
-        'ibm866',       '866',          'cp1251',       'windows-1251',
-        'win-1251',     '1251',         'cp1252',       'windows-1252',
-        '1252',         'koi8-r',       'koi8-ru',      'koi8r',
-        'big5',         '950',          'gb2312',       '936',
-        'big5-hkscs',   'shift_jis',    'sjis',         'sjis-win',
-        'cp932',        '932',          'euc-jp',       'eucjp',
-        'eucjp-win',    'macroman'
-    ];
+    use EscaperEncodingsTrait;
 
     /** @var EscapeHelper */
     private $helper;
@@ -34,14 +24,6 @@ class EscapeCssTest extends TestCase
     protected function setUp(): void
     {
         $this->helper = new EscapeHelper;
-    }
-
-    /** @return iterable<string, array<int, string>> */
-    public function supportedEncodingsProvider() : iterable
-    {
-        foreach ($this->supportedEncodings as $encoding) {
-            yield $encoding => [$encoding];
-        }
     }
 
     public function testUsesUtf8EncodingByDefault()
